Moved the uzip method from modelUtils to fileUtils
