+++ utils/src/main/resources/config.properties
@@ -2,6 +2,7 @@ name=Dynamic-Analyser
 mode=SOURCECODE
 trace=queue
 traceFile=trace
+indexDirectory=index
 #AUTO will use maven to compute the dependencies
 spoon.mode=AUTO
 sutBinaries=/target/classes
